requirements/requirements_docs.txt
sphinx>=3.1.2
sphinx-bootstrap-theme>=0.6.5
sphinx-automodapi
numpy>=1.12.1
pyzmq>=17.1.2
# pyo>=0.7.8
pandas>=0.19.2
npyscreen
scipy
tornado
pigpio
git+https://github.com/pyqtgraph/pyqtgraph@develop
inputs
altair
scikit-learn
JACK-Client
tqdm
inputs
bokeh
colorcet
PySide2
autodocsumm
sphinx_rtd_theme
tables
blosc
scikit-video
sphinx-autobuild
git+https://github.com/mwibrow/sphinx-sass.git